+++ src/module/settings.ts
@@ -26,11 +26,27 @@ export interface StatusEffect {
 	return canvas;
 }
 
+/**
+ * Because typescript doesn't know when in the lifecycle of foundry your code runs, we have to assume that the
+ * canvas is potentially not yet initialized, so it's typed as declare let canvas: Canvas | {ready: false}.
+ * That's why you get errors when you try to access properties on canvas other than ready.
+ * In order to get around that, you need to type guard canvas.
+ * Also be aware that this will become even more important in 0.8.x because no canvas mode is being introduced there.
+ * So you will need to deal with the fact that there might not be an initialized canvas at any point in time.
+ * @returns
+ */
+ export function getGame(): Game {
+	if (!(game instanceof Game)) {
+		throw new Error("Game Is Not Initialized");
+	}
+	return game;
+}
+
 export const registerSettings = function() {
 
-	game.settings.register(MODULE_NAME, "autoStealth", {
-		name: game.i18n.localize("conditional-visibility.settings.autoStealth.name"),
-		hint: game.i18n.localize("conditional-visibility.settings.autoStealth.hint"),
+	getGame().settings.register(MODULE_NAME, "autoStealth", {
+		name: getGame().i18n.localize("conditional-visibility.settings.autoStealth.name"),
+		hint: getGame().i18n.localize("conditional-visibility.settings.autoStealth.hint"),
 		scope: "world",
 		config: true,
 		type: Boolean,
@@ -38,13 +54,13 @@ export const registerSettings = function() {
 		onChange: value => console.log(MODULE_NAME + ' | autoStealth set to ' + value)
 	});
 
-	game.settings.register(MODULE_NAME, "popup-version", {
+	getGame().settings.register(MODULE_NAME, "popup-version", {
 		scope: "world",
 		config: false,
 		type: String,
 		default: "0.0.9",
 	  });
-	  
+
 	// Register any custom module settings here
-	ConditionalVisibility.onInit();	
+	ConditionalVisibility.onInit();
 }